+++ native/scripts/build-release-gl64.sh
@@ -1,6 +1,11 @@
 #!/bin/sh
 
-# CAP=`./configure.sh | grep capability | cut -d ' ' -f 3`
+if ! [ -d "scripts" ]; then
+    echo -e "Error: scripts directory not found. Please run this script from the \e[44mnative\e[0m directory of the repository."
+    exit 1
+fi
+
+# CAP=`./scripts/configure.sh | grep capability | cut -d ' ' -f 3`
 CAP="86;89;90"
 
 rm -rf build
